requirements.txt
openai
python-dotenv
pydantic
streamlit
pandas
datasets
